+++ src/h264bsd_image.c
@@ -1,6 +1,5 @@
 /*
  * Copyright (C) 2009 The Android Open Source Project
- * Modified for use by h264bsd standalone library
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -93,7 +92,7 @@ void h264bsdWriteMacroblock(image_t *image, u8 *data)
 
     ASSERT(image);
     ASSERT(data);
-    ASSERT(!((size_t)data&0x3));
+    ASSERT(!((u32)data&0x3));
 
     width = image->width;
 
@@ -102,9 +101,9 @@ void h264bsdWriteMacroblock(image_t *image, u8 *data)
     lum = (u32*)image->luma;
     cb = (u32*)image->cb;
     cr = (u32*)image->cr;
-    ASSERT(!((size_t)lum&0x3));
-    ASSERT(!((size_t)cb&0x3));
-    ASSERT(!((size_t)cr&0x3));
+    ASSERT(!((u32)lum&0x3));
+    ASSERT(!((u32)cb&0x3));
+    ASSERT(!((u32)cr&0x3));
 
     ptr = (u32*)data;
 
@@ -192,7 +191,7 @@ void h264bsdWriteOutputBlocks(image_t *image, u32 mbNum, u8 *data,
     ASSERT(image);
     ASSERT(data);
     ASSERT(mbNum < image->width * image->height);
-    ASSERT(!((size_t)data&0x3));
+    ASSERT(!((u32)data&0x3));
 
     /* Image size in macroblocks */
     picWidth = image->width;
@@ -219,8 +218,8 @@ void h264bsdWriteOutputBlocks(image_t *image, u32 mbNum, u8 *data,
         tmp = data + y*16 + x;
         imageBlock = lum + y*picWidth + x;
 
-        ASSERT(!((size_t)tmp&0x3));
-        ASSERT(!((size_t)imageBlock&0x3));
+        ASSERT(!((u32)tmp&0x3));
+        ASSERT(!((u32)imageBlock&0x3));
 
         if (IS_RESIDUAL_EMPTY(pRes))
         {
@@ -293,8 +292,8 @@ void h264bsdWriteOutputBlocks(image_t *image, u32 mbNum, u8 *data,
         tmp += y*8 + x;
         imageBlock += y*picWidth + x;
 
-        ASSERT(!((size_t)tmp&0x3));
-        ASSERT(!((size_t)imageBlock&0x3));
+        ASSERT(!((u32)tmp&0x3));
+        ASSERT(!((u32)imageBlock&0x3));
 
         if (IS_RESIDUAL_EMPTY(pRes))
         {
